refactor: List hook, component type 변환#76
