+++ src/components/HistoryList.tsx
@@ -1,7 +1,6 @@
-
-import { HistoryData } from "@/types/ListType";
-import ListTableComponent from "@/components/ListTableComponent";
+import { HistoryData, HistoryDataTypeValue } from "@/types/ListType";
 import { formatDate } from "@/utils/formatDate";
+import ListTableComponent from "@/components/ListTableComponent";
 
 interface HistoryListProp {
   userHistoryData: HistoryData[];
@@ -18,14 +17,14 @@ const HistoryList = ({
     {
       key: "startDate",
       label: "대여일",
-      render: (value: string | null) =>
-        value ? formatDate(value) : "날짜 정보를 불러올 수 없습니다.",
+      render: (value: HistoryDataTypeValue) =>
+        value ? formatDate(value as string) : "날짜 정보를 불러올 수 없습니다.",
     },
     {
       key: "endDate",
       label: "반납일",
-      render: (value: string | null) =>
-        value ? formatDate(value) : "날짜 정보를 불러올 수 없습니다.",
+      render: (value: HistoryDataTypeValue) =>
+        value ? formatDate(value as string) : "날짜 정보를 불러올 수 없습니다.",
     },
   ];
   return (
